refactor(Types.Code):  to Types.CodeMirror to make it easier to add support for the many other features CodeMirror supports.
