+++ admin/public/styles/keystone/forms.less
@@ -58,10 +58,10 @@
 	}
 }
 
-// Code
+// CodeMirror
 /*
 -- TODO @jossmac: this update from the master branch may need to be brought forward
-.field-type-code {
+.field-type-codemirror {
 	.CodeMirror-container {
 		.form-control;
 		height: auto;
