+++ fields/types/codemirror/CodeMirrorType.js
@@ -8,20 +8,20 @@ var util = require('util');
  * @extends Field
  * @api public
  */
-function code(list, path, options) {
+function codemirror(list, path, options) {
 	this._nativeType = String;
 	this._defaultSize = 'full';
 	this.height = options.height || 180;
 	this.lang = options.lang || options.language;
 	this._properties = ['editor', 'height', 'lang'];
-	this.codemirror = options.codemirror || {};
-	this.editor = _.defaults(this.codemirror, { mode : this.lang });
-	code.super_.call(this, list, path, options);
+	this.config = options.config || {};
+	this.editor = _.defaults(this.config, { mode : this.lang });
+	codemirror.super_.call(this, list, path, options);
 }
-util.inherits(code, FieldType);
+util.inherits(codemirror, FieldType);
 
 /* Inherit from TextType prototype */
-code.prototype.addFilterToQuery = TextType.prototype.addFilterToQuery;
+codemirror.prototype.addFilterToQuery = TextType.prototype.addFilterToQuery;
 
 /* Export Field Type */
-module.exports = code;
+module.exports = codemirror;
